+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+- Add review summary [(#87)](https://github.com/plotwist-app/plotwist-backend/pull/87)
+
 ## 1.9.0
 
 ### Added